% Auto-generated: do not edit by hand
\name{''Micro}

\alias{''Micro}

\title{Micro component}

\description{

}

\usage{
''Micro(id=NULL, fileUrl=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{fileUrl}{Character. The URL for the temp audio file if you saved it to the server.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

