% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{addBatch_EL4P}
\alias{addBatch_EL4P}
\title{EL4P Add Egg Batch}
\usage{
addBatch_EL4P(batch)
}
\arguments{
\item{batch}{numeric vector, size of vector must equal the number of genotypes in the EL4P population}
}
\description{
Add an egg batch to EL4P
\itemize{
\item This method is bound to \code{EL4P$addBatch_EL4P}
}
}
