% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-PeriDomestic.R
\name{clear_EggQ_periDomestic_AquaticSite}
\alias{clear_EggQ_periDomestic_AquaticSite}
\title{Clear EggQ}
\usage{
clear_EggQ_periDomestic_AquaticSite()
}
\description{
Clear the egg queue at this site, an object of class \code{\link[MASHcpp]{EggQ}}
\itemize{
\item This method is bound to \code{periDomestic_AquaticSite$clear_EggQ}
}
}
