% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{clear_lambda_EL4P}
\alias{clear_lambda_EL4P}
\title{EL4P: Clear Lambda}
\usage{
clear_lambda_EL4P()
}
\description{
Clear lambda
\itemize{
\item This method is bound to \code{EL4P$clear_lambda}
}
}
