% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{clear_MatingQ_Landscape}
\alias{clear_MatingQ_Landscape}
\title{Clear Mating Queue for Landscape}
\usage{
clear_MatingQ_Landscape()
}
\description{
Clear the \code{\link[MASHcpp]{MatingQ}} for all mating sites by calling \code{\link{clear_MatingQ_MatingSite}}
This function is called in \code{\link{mbitesMale_Pop_MBITES}}.
\itemize{
\item This method is bound to \code{Landscape$clear_MatingQ}
}
}
