% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{clear_MatingQ_MatingSite}
\alias{clear_MatingQ_MatingSite}
\title{Clear MatingQ}
\usage{
clear_MatingQ_MatingSite()
}
\description{
Clear the mating queue object at this site; see \code{\link[MASHcpp]{MatingQ}} for details.
\itemize{
\item This method is bound to \code{MatingSite$clear_MatingQ}
}
}
