% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{clear_RiskQ_FeedingSite}
\alias{clear_RiskQ_FeedingSite}
\title{Clear RiskQ}
\usage{
clear_RiskQ_FeedingSite()
}
\description{
Clear the human risk queue at this site, an object of class \code{\link[MASHcpp]{RiskQ}}
\itemize{
\item This method is bound to \code{FeedingSite$clear_RiskQ}
}
}
