% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{clear_RiskQ_Landscape}
\alias{clear_RiskQ_Landscape}
\title{Clear Host Risk Queue for Landscape}
\usage{
clear_RiskQ_Landscape()
}
\description{
Clear the \code{\link[MASHcpp]{RiskQ}} for all blood feeding sites by calling \code{\link{clear_RiskQ_FeedingSite}}
\itemize{
\item This method is bound to \code{Landscape$clear_RiskQ}
}
}
