% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{close_HumanPathogenCon_Tile}
\alias{close_HumanPathogenCon_Tile}
\title{Close Human Pathogen JSON Connection}
\usage{
close_HumanPathogenCon_Tile()
}
\description{
Close the text connection used to write human-stage pathogen data to JSON
\itemize{
\item This method is bound to \code{Tile$close_HumanPathogenCon}
}
}
