% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{close_MosquitoPathogenCon_Tile}
\alias{close_MosquitoPathogenCon_Tile}
\title{Close Mosquito Pathogen JSON Connection}
\usage{
close_MosquitoPathogenCon_Tile()
}
\description{
Close the text connection used to write mosquito-stage pathogen data to JSON
\itemize{
\item This method is bound to \code{Tile$close_MosquitoPathogenCon}
}
}
