% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_diffs.R
\name{column_diffs}
\alias{column_diffs}
\title{Calculate column differences}
\usage{
column_diffs(
  data,
  prefix.pattern = NULL,
  include.first = TRUE,
  suffix.out = "_diff"
)
}
\arguments{
\item{data}{tibble}

\item{prefix.pattern}{tidyselect style column selection. Default is NULL}

\item{include.first}{include first column. Default is TRUE}

\item{suffix.out}{suffix for output columns}
}
\value{
tibble
}
\description{
Calculate column differences
}
\examples{
tibble::tibble(a=sample(1:10,5),b=sample(1:10,5),c=sample(1:10,5)) |>
(\(.x) list(.x,column_diffs(.x)))()

}
