% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEBUG.R
\name{DEBUG.MASHMICRO}
\alias{DEBUG.MASHMICRO}
\title{Debug Finalizers for Classes}
\usage{
DEBUG.MASHMICRO(MASHCPP = FALSE, overwrite = TRUE)
}
\description{
Print when objects are garbage collected. For debugging purposes only.
}
