% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Class.R
\docType{class}
\name{EL4P}
\alias{EL4P}
\title{EL4P Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
EL4P
}
\description{
im a class!
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
