% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Fit.R
\name{EL4P_fitK}
\alias{EL4P_fitK}
\title{EL4P Fit K}
\usage{
EL4P_fitK(lambda, beta = 32, mu_E = 0.168, mu_L = 0.168, mu_P = 0.168,
  mu_M = 0.123, d_E = 1, d_L = 14, d_P = 1)
}
\arguments{
\item{beta}{daily rate of oviposition of eggs (units in 1/day * 1/female mosquito)}

\item{mu_E}{egg stage daily mortality rate}

\item{mu_L}{larvae stage daily mortality rate}

\item{mu_P}{pupae stage daily mortality rate}

\item{mu_M}{adult female daily mortality rate}

\item{d_E}{mean duration of egg stage (days)}

\item{d_L}{mean duration of larvae stage (days)}

\item{d_P}{mean duration of pupae stage (days)}
}
\description{
Given equilibrium parameters and emergence lambda, fit carrying capacity K.
}
