% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-ActivitySpace-Methods.R
\name{get_ActivitySpace_Tile}
\alias{get_ActivitySpace_Tile}
\title{Return Human Activity Space on a Tile}
\usage{
get_ActivitySpace_Tile()
}
\description{
Return the human activity space for each site.
\itemize{
\item This function is bound to \code{Tile$get_ActivitySpace()}
}
}
