% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{get_all_EL4P}
\alias{get_all_EL4P}
\title{EL4P: Get All Life Stages}
\usage{
get_all_EL4P(lag = FALSE)
}
\arguments{
\item{lag}{logical; return lag stages as well?}
}
\description{
Return all life stages
\itemize{
\item This method is bound to \code{EL4P$get_all}
}
}
