% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_AquaSitesN_Landscape}
\alias{get_AquaSitesN_Landscape}
\title{Get Number of Aquatic Habitat Sites}
\usage{
get_AquaSitesN_Landscape()
}
\description{
\itemize{
\item This method is bound to \code{Landscape$get_AquaSitesN}
}
}
