% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_bDay_Mosquito}
\alias{get_bDay_Mosquito}
\title{Get bDay}
\usage{
get_bDay_Mosquito()
}
\description{
Return mosquito birthday (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_bDay} and \code{MosquitoMale$get_bDay}
}
}
