% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods-VectorControl.R
\name{get_biologicalControl}
\alias{get_biologicalControl}
\title{Get swarmSpray}
\usage{
get_biologicalControl()
}
\arguments{
\item{swarmSpray}{Object}
}
\description{
Return the definition of the Swarm Spray
}
