% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_bionomics_Mosquito}
\alias{get_bionomics_Mosquito}
\title{Get Bionomics}
\usage{
get_bionomics_Mosquito()
}
\description{
Return bionomics calculated from \code{\link[MASHcpp]{MosquitoFemaleHistory}}
\itemize{
\item This method is bound to \code{MosquitoFemale$get_bionomics}
}
}
