% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_damage_Mosquito}
\alias{get_damage_Mosquito}
\title{Get damage}
\usage{
get_damage_Mosquito()
}
\description{
Return mosquito wing damage (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_damage} and \code{MosquitoMale$get_damage}
}
}
