% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_eggP_Mosquito}
\alias{get_eggP_Mosquito}
\title{Get eggP}
\usage{
get_eggP_Mosquito()
}
\description{
Return mosquito minimum blood provision for eggs to mature (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_eggP}
}
}
