% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_EggQ_AquaticSite}
\alias{get_EggQ_AquaticSite}
\title{Get EggQ}
\usage{
get_EggQ_AquaticSite()
}
\description{
Get mosquito egg batch queue, object of class \code{\link[MASHcpp]{EggQ}}
\itemize{
\item This method is bound to \code{AquaticSite$get_EggQ}
}
}
