% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_eggT_Mosquito}
\alias{get_eggT_Mosquito}
\title{Get eggT}
\usage{
get_eggT_Mosquito()
}
\description{
Return mosquito minimum time for eggs to mature (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_eggT}
}
}
