% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-PeriDomestic.R
\name{get_EL4P_periDomestic_AquaticSite}
\alias{get_EL4P_periDomestic_AquaticSite}
\title{Get EL4P}
\usage{
get_EL4P_periDomestic_AquaticSite()
}
\description{
Get aquatic populations, object of class \code{\link[MASHcpp]{EL4P}}
\itemize{
\item This method is bound to \code{periDomestic_AquaticSite$get_EL4P}
}
}
