% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_energyPreG_Mosquito}
\alias{get_energyPreG_Mosquito}
\title{Get energyPreG}
\usage{
get_energyPreG_Mosquito()
}
\description{
Return mosquito minimum blood provision for maturation (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_energyPreG}
}
}
