% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_enterP_FeedingSite}
\alias{get_enterP_FeedingSite}
\title{Get enterP}
\usage{
get_enterP_FeedingSite()
}
\description{
Get house entry probability at this site \code{enterP}
\itemize{
\item This method is bound to \code{FeedingSite$get_enterP}
}
}
