% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_FemaleCSVCon_Tile}
\alias{get_FemaleCSVCon_Tile}
\title{Get Female CSV Connection}
\usage{
get_FemaleCSVCon_Tile()
}
\description{
Return the text connection used to write population count data to CSV.
\itemize{
\item This method is bound to \code{Tile$get_FemaleCSVCon}
}
}
