% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_FemaleHistoryCon_Tile}
\alias{get_FemaleHistoryCon_Tile}
\title{Get Female JSON Connection}
\usage{
get_FemaleHistoryCon_Tile()
}
\description{
Return the text connection used to write female histories to JSON.
\itemize{
\item This method is bound to \code{Tile$get_FemaleHistoryCon}
}
}
