% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Setup.R
\name{get_FemalePopPointer_Human}
\alias{get_FemalePopPointer_Human}
\title{Get \code{\link{MosquitoPopFemale}} Pointer}
\usage{
get_FemalePopPointer_Human()
}
\description{
Return pointer to the female mosquito population on this landscape.
\itemize{
\item This method is bound to \code{Human$get_FemalePopPointer}
}
}
