% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_FemalePopPointer_Landscape}
\alias{get_FemalePopPointer_Landscape}
\title{Get \code{\link{MosquitoPopFemale}} Pointer}
\usage{
get_FemalePopPointer_Landscape()
}
\description{
Return pointer to \code{MosquitoPopFemale$self}.
\itemize{
\item This method is bound to \code{Landscape$get_FemalePopPointer}
}
}
