% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_genotype_Mosquito}
\alias{get_genotype_Mosquito}
\title{Get genotype}
\usage{
get_genotype_Mosquito()
}
\description{
Return mosquito genotype (\code{integer}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_genotype} and \code{MosquitoMale$get_genotype}
}
}
