% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_hazLspot_FeedingSite}
\alias{get_hazLspot_FeedingSite}
\title{Get Landing Hazards}
\usage{
get_hazLspot_FeedingSite(lspot)
}
\description{
Get landing hazard associated with one of the spots a mosquito may land at for this site.
\itemize{
\item This method is bound to \code{FeedingSite$get_hazLspot}
}
}
