% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_hazV_FeedingSite}
\alias{get_hazV_FeedingSite}
\title{Get hazV}
\usage{
get_hazV_FeedingSite()
}
\description{
Get landing hazard associated with vegetation at this site \code{hazV}
\itemize{
\item This method is bound to \code{FeedingSite$get_hazV}
}
}
