% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_hazW_FeedingSite}
\alias{get_hazW_FeedingSite}
\title{Get hazW}
\usage{
get_hazW_FeedingSite()
}
\description{
Get landing hazard associated with outside wall at this site \code{get_hazW}
\itemize{
\item This method is bound to \code{FeedingSite$get_hazW}
}
}
