% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_history_Mosquito}
\alias{get_history_Mosquito}
\title{Get History}
\usage{
get_history_Mosquito()
}
\description{
Return history from \code{\link[MASHcpp]{MosquitoFemaleHistory}}
\itemize{
\item This method is bound to \code{MosquitoFemale$get_history}
}
}
