% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_hostID_Mosquito}
\alias{get_hostID_Mosquito}
\title{Get hostID}
\usage{
get_hostID_Mosquito()
}
\description{
Return ID of the host this mosquito last fed upon (\code{integer}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_hostID}
}
}
