% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_HumanPop_Tile}
\alias{get_HumanPop_Tile}
\title{Get Humans}
\usage{
get_HumanPop_Tile()
}
\description{
Return embedded human population object \code{\link[MASHmacro]{HumanPop}}
\itemize{
\item This method is bound to \code{Tile$get_HumanPop}
}
}
