% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_HumansPointer_Mosquito}
\alias{get_HumansPointer_Mosquito}
\title{Get \code{\link[MASHmacro]{HumanPop}} Pointer}
\usage{
get_HumansPointer_Mosquito()
}
\description{
Return pointer to \code{HumanPop$self} within the same microsimulation tile by reference.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_HumansPointer} and \code{MosquitoMale$get_HumansPointer}
}
}
