% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{get_HumansPointer_MosquitoPop}
\alias{get_HumansPointer_MosquitoPop}
\title{Get \code{\link[MASHmacro]{HumanPop}} Pointer}
\usage{
get_HumansPointer_MosquitoPop()
}
\description{
Return pointer to \code{HumanPop$self} within the same microsimulation tile by reference.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$get_HumansPointer} and \code{MosquitoPopMale$get_HumansPointer}
}
}
