% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_id_Mosquito}
\alias{get_id_Mosquito}
\title{Get ID}
\usage{
get_id_Mosquito()
}
\description{
Return mosquito ID (\code{character}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_id} and \code{MosquitoMale$get_id}
}
}
