% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_ImagoQ_AquaticSite}
\alias{get_ImagoQ_AquaticSite}
\title{Get ImagoQ}
\usage{
get_ImagoQ_AquaticSite()
}
\description{
Get mosquito imago queue of emerging adults, object of class \code{\link[MASHcpp]{ImagoQ}}
\itemize{
\item This method is bound to \code{AquaticSite$get_ImagoQ}
}
}
