% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_ix_AquaticSite}
\alias{get_ix_AquaticSite}
\title{Get ix}
\usage{
get_ix_AquaticSite()
}
\description{
Return site index \code{ix}
\itemize{
\item This method is bound to \code{AquaticSite$get_ix}
}
}
