% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_lambda_AquaticSite}
\alias{get_lambda_AquaticSite}
\title{Get lambda}
\usage{
get_lambda_AquaticSite()
}
\description{
Get vector of lambda of mean daily emergence at this site
\itemize{
\item This method is bound to \code{AquaticSite$get_lambda}
}
}
