% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Setup.R
\name{get_LandscapePointer_Human}
\alias{get_LandscapePointer_Human}
\title{Get \code{\link{Landscape}} Pointer}
\usage{
get_LandscapePointer_Human()
}
\description{
Return pointer to \code{Landscape$self} this population lives in.
\itemize{
\item This method is bound to \code{Human$get_LandscapePointer}
}
}
