% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-PeriDomestic.R
\name{get_LandscapePointer_periDomestic_AquaticSite}
\alias{get_LandscapePointer_periDomestic_AquaticSite}
\title{Get \code{\link{Landscape}} Pointer}
\usage{
get_LandscapePointer_periDomestic_AquaticSite()
}
\description{
Return \code{Landscape$self} enclosing this site.
\itemize{
\item This method is bound to \code{periDomestic_AquaticSite$get_LandscapePointer}
}
}
