% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_LandscapePointer_SugarSite}
\alias{get_LandscapePointer_SugarSite}
\title{Get \code{\link{Landscape}} Pointer}
\usage{
get_LandscapePointer_SugarSite()
}
\description{
Return \code{Landscape$self} enclosing this site.
\itemize{
\item This method is bound to \code{SugarSite$get_LandscapePointer}
}
}
