% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_locNow_Mosquito}
\alias{get_locNow_Mosquito}
\title{Get locNow}
\usage{
get_locNow_Mosquito()
}
\description{
Return next mosquito point set (\code{character}). With \code{pSetNow}, the next site of the mosquito's spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_locNow} and \code{MosquitoMale$get_locNow}
}
}
