% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_locOld_Mosquito}
\alias{get_locOld_Mosquito}
\title{Get locOld}
\usage{
get_locOld_Mosquito()
}
\description{
Return next mosquito point set (\code{character}). With \code{pSetOld}, the next site of the mosquito's spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_locOld} and \code{MosquitoMale$get_locOld}
}
}
