% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_lspot_Mosquito}
\alias{get_lspot_Mosquito}
\title{Get lspot}
\usage{
get_lspot_Mosquito()
}
\description{
Return mosquito landing spot (\code{character}).
Codes corresponding to the following spatial relation to the site this mosquito is at.
\itemize{
\item i: rest on the inside wall of a structure
\item w: rest on the outside wall of a structure
\item v: rest on vegetation
\item r: reattempt without resting
\item l: leave the area
}
}
\details{
\itemize{
\item This method is bound to \code{MosquitoFemale$get_lspot} and \code{MosquitoMale$get_lspot}
}
}
