% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_MaleCSVCon_Tile}
\alias{get_MaleCSVCon_Tile}
\title{Get Male CSV Connection}
\usage{
get_MaleCSVCon_Tile()
}
\description{
Return the text connection used to write population count data to CSV.
\itemize{
\item This method is bound to \code{Tile$get_MaleCSVCon}
}
}
