% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_MalePopPointer_Mosquito}
\alias{get_MalePopPointer_Mosquito}
\title{Get \code{\link{MosquitoPopMale}} Pointer}
\usage{
get_MalePopPointer_Mosquito()
}
\description{
Return \code{MosquitoPopMale$self} enclosing this mosquito (if method for \code{\link{MosquitoMale}}), or in the same \code{\link{Tile}} (if method for \code{\link{MosquitoFemale}}) by reference.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_MalePopPointer} and \code{MosquitoMale$get_MalePopPointer}
}
}
