% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{get_mateFitness_MosquitoPopMale}
\alias{get_mateFitness_MosquitoPopMale}
\title{Get mateFitness}
\usage{
get_mateFitness_MosquitoPopMale(genotype)
}
\arguments{
\item{genotype}{integer; genotype of male mosquito}
}
\description{
Get the male mating fitness associated with a given genotype.
\itemize{
\item This method is bound to \code{MosquitoPopMale$get_mateFitness}
}
}
