% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_mateGenotype_Mosquito}
\alias{get_mateGenotype_Mosquito}
\title{Get mateGenotype}
\usage{
get_mateGenotype_Mosquito()
}
\description{
Return genotype of the \code{\link{MosquitoMale}} this mosquito mated with.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_mateGenotype}
}
}
