% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_mateID_Mosquito}
\alias{get_mateID_Mosquito}
\title{Get mateID}
\usage{
get_mateID_Mosquito()
}
\description{
Return ID of the \code{\link{MosquitoMale}} this mosquito mated with.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_mateID}
}
}
