% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_MatingQ_MatingSite}
\alias{get_MatingQ_MatingSite}
\title{Get MatingQ}
\usage{
get_MatingQ_MatingSite()
}
\description{
Get the mating queue object at this site; see \code{\link[MASHcpp]{MatingQ}} for details.
\itemize{
\item This method is bound to \code{MatingSite$get_MatingQ}
}
}
