% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_MatingSites_Landscape}
\alias{get_MatingSites_Landscape}
\title{Get Mating Sites}
\usage{
get_MatingSites_Landscape(ixS = NULL)
}
\arguments{
\item{ixS}{integer index; if \code{NULL} return entire list of sites}
}
\description{
Return single \code{\link{MatingSite}} or all sites.
\itemize{
\item This method is bound to \code{Landscape$get_MatingSites}
}
}
