% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{get_MBITES_PAR_MosquitoPop}
\alias{get_MBITES_PAR_MosquitoPop}
\title{Get MBITES_PAR}
\usage{
get_MBITES_PAR_MosquitoPop(par = NULL)
}
\arguments{
\item{par}{if given, return named parameter otherwise return entire list}
}
\description{
Get an M-BITES parameter.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$get_MBITES_PAR} and \code{MosquitoPopFemale$get_MBITES_PAR}
}
}
