% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_MosquitoPathogenCon_Tile}
\alias{get_MosquitoPathogenCon_Tile}
\title{Get Mosquito Pathogen JSON Connection}
\usage{
get_MosquitoPathogenCon_Tile()
}
\description{
Return the text connection used to write mosquito-stage pathogen data to JSON
\itemize{
\item This method is bound to \code{Tile$get_MosquitoPathogenCon}
}
}
