% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{get_N_MosquitoPop}
\alias{get_N_MosquitoPop}
\title{Get Number of Alive Mosquitoes}
\usage{
get_N_MosquitoPop()
}
\description{
Provide the number of living mosquitoes.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$get_N} and \code{MosquitoPopFemale$get_N}
}
}
