% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_Pathogens_Mosquito}
\alias{get_Pathogens_Mosquito}
\title{Get Pathogens}
\usage{
get_Pathogens_Mosquito()
}
\description{
Return \code{Pathogens} object in this mosquito.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_Pathogens}
}
}
