% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{get_pop_MosquitoPop}
\alias{get_pop_MosquitoPop}
\title{Get pop}
\usage{
get_pop_MosquitoPop()
}
\description{
Simply provides access to the \code{\link[MASHcpp]{HashMap}} object that contains the mosquito objects.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$get_pop} and \code{MosquitoPopFemale$get_pop}
}
}
