% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_pSetOld_Mosquito}
\alias{get_pSetOld_Mosquito}
\title{Get pSetOld}
\usage{
get_pSetOld_Mosquito()
}
\description{
Return next mosquito point set (\code{character}). With \code{locOld}, the next site of the mosquito's spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_pSetOld} and \code{MosquitoMale$get_pSetOld}
}
}
