% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_RiskQ_FeedingSite}
\alias{get_RiskQ_FeedingSite}
\title{Get RiskQ}
\usage{
get_RiskQ_FeedingSite()
}
\description{
Get the host risk queue at this site, an object of class \code{\link[MASHcpp]{RiskQ}}
\itemize{
\item This method is bound to \code{FeedingSite$get_RiskQ}
}
}
