% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_runID_Tile}
\alias{get_runID_Tile}
\title{Get Run ID}
\usage{
get_runID_Tile()
}
\description{
Return current simulation run ID that will be used to identify multiple simulation output from the same tile object.
\itemize{
\item This method is bound to \code{Tile$get_runID}
}
}
