% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_searchWt_MatingSite}
\alias{get_searchWt_MatingSite}
\title{Get searchWt}
\usage{
get_searchWt_MatingSite()
}
\description{
Get site search weight \code{searchWt}
\itemize{
\item This method is bound to \code{MatingSite$get_searchWt}
}
}
