% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_siteXY_SugarSite}
\alias{get_siteXY_SugarSite}
\title{Get siteXY}
\usage{
get_siteXY_SugarSite()
}
\description{
Get site coordinates \code{siteXY}
\itemize{
\item This method is bound to \code{SugarSite$get_siteXY}
}
}
