% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_state_Mosquito}
\alias{get_state_Mosquito}
\title{Get state}
\usage{
get_state_Mosquito()
}
\description{
Return mosquito current behavioral state (\code{character}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_state} and \code{MosquitoMale$get_state}
}
}
