% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_stateNew_Mosquito}
\alias{get_stateNew_Mosquito}
\title{Get stateNew}
\usage{
get_stateNew_Mosquito()
}
\description{
Return mosquito next behavioral state (\code{character}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_stateNew} and \code{MosquitoMale$get_stateNew}
}
}
