% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_SugarSites_Landscape}
\alias{get_SugarSites_Landscape}
\title{Get Sugar Feeding Sites}
\usage{
get_SugarSites_Landscape(ixS = NULL)
}
\arguments{
\item{ixS}{integer index; if \code{NULL} return entire list of sites}
}
\description{
Return single \code{\link{SugarSite}} or all sites.
\itemize{
\item This method is bound to \code{Landscape$get_SugarSites}
}
}
