% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_SugarSitesN_Landscape}
\alias{get_SugarSitesN_Landscape}
\title{Get Number of Sugar Feeding Sites}
\usage{
get_SugarSitesN_Landscape()
}
\description{
\itemize{
\item This method is bound to \code{Landscape$get_SugarSitesN}
}
}
