% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_TilePointer_Mosquito}
\alias{get_TilePointer_Mosquito}
\title{Get \code{\link{Tile}} Pointer}
\usage{
get_TilePointer_Mosquito()
}
\description{
Return pointer to enclosing \code{Tile$self}.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_TilePointer} and \code{MosquitoMale$get_TilePointer}
}
}
