% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_tNext_Mosquito}
\alias{get_tNext_Mosquito}
\title{Get tNext}
\usage{
get_tNext_Mosquito()
}
\description{
Return mosquito next jump time (time at which mosquito will leave current state) (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_tNext} and \code{MosquitoMale$get_tNext}
}
}
