% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_tNow_Tile}
\alias{get_tNow_Tile}
\title{Get tNow}
\usage{
get_tNow_Tile()
}
\description{
Return current simulation time. All agents run events until their next scheduled event overruns the current simulation time, upon which they
hold in the current behavioral state until the next time step. At the end of all activity on that iteration, \code{tNow} increases and the loop iterates.
\itemize{
\item This method is bound to \code{Tile$get_tNow}
}
}
