% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-ActivitySpace-Methods.R
\name{init_ActivitySpace_Human}
\alias{init_ActivitySpace_Human}
\title{Initialize Activity Space for \code{\link[MASHmacro]{Human}}}
\usage{
init_ActivitySpace_Human(nDaily)
}
\arguments{
\item{nDaily}{average daily number of other sites visited}
}
\description{
This function is bound to \code{Human$init_ActivitySpace()}
}
