% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-ActivitySpace-Methods.R
\name{init_ActivitySpace_HumanPop}
\alias{init_ActivitySpace_HumanPop}
\title{Initialize Activity Space for \code{\link[MASHmacro]{HumanPop}}}
\usage{
init_ActivitySpace_HumanPop(nDaily = 1.4)
}
\arguments{
\item{nDaily}{average daily number of other sites visited}
}
\description{
This function is bound to \code{HumanPop$init_ActivitySpace()}
}
