% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{init_Pathogens_Mosquito}
\alias{init_Pathogens_Mosquito}
\title{Initialize Pathogens}
\usage{
init_Pathogens_Mosquito()
}
\description{
Initialize \code{Pathogens} object in this mosquito. This function is a placeholder that should be overwritten by an appropriate PATHOGEN module setup function.
\itemize{
\item This method is bound to \code{MosquitoFemale$init_Pathogens}
}
}
