% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{init_Pathogens_MosquitoFemale_PfSI}
\alias{init_Pathogens_MosquitoFemale_PfSI}
\title{PfSI Helper Code for Pathogen Initialization in \code{\link{MosquitoFemale}}}
\usage{
init_Pathogens_MosquitoFemale_PfSI()
}
\description{
Initializes an empty PfSI pathogen in \code{\link{MosquitoFemale}} called during object initialization.
\itemize{
\item This method is bound to \code{MosquitoFemale$init_Pathogens()}
}
}
