% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Methods.R
\name{init_Pathogens_MosquitoPopFemale_PfSI}
\alias{init_Pathogens_MosquitoPopFemale_PfSI}
\title{PfSI Helper Code for Pathogen Initialization in \code{\link{MosquitoPopFemale}}}
\usage{
init_Pathogens_MosquitoPopFemale_PfSI()
}
\description{
Initializes empty PfSI pathogens in \code{\link{MosquitoPopFemale}} called during object initialization.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$init_Pathogens()}
}
}
