% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Parameters.R
\name{Landscape.Sugar.Parameters}
\alias{Landscape.Sugar.Parameters}
\title{MICRO: Generate Parameters for \code{\link{Landscape}} \code{\link{SugarSite}}}
\usage{
Landscape.Sugar.Parameters(nSugar, siteXY, searchWt = NULL, haz = 0)
}
\arguments{
\item{nSugar}{number of sugar feeding sites}

\item{siteXY}{two element list of \code{x} and \code{y} coordinates of mating sites}

\item{searchWt}{vector of searchWt (if \code{NULL} initialize to Gamma(1,1) distribution)}

\item{haz}{mean value of landing hazards (if \code{!= 0} use \code{\link{betaRootA}} to find alpha parameter of beta distribution to give that mean value and produce Beta distributed hazards)}
}
\value{
return a list
}
\description{
This function generates a named list of parameters to initialize all \code{\link{SugarSite}} objects on a MICRO \code{\link{Landscape}}.
}
