% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_micro_table.R
\name{mask_micro_table}
\alias{mask_micro_table}
\title{Mask across tibble by column differences}
\usage{
mask_micro_table(data, micro.n = 5, down = TRUE, col.sel)
}
\arguments{
\item{data}{tibble}

\item{micro.n}{n to mask at}

\item{down}{round down in interval}

\item{col.sel}{tidyselect style column selection for masking. Required.}
}
\value{
tibble
}
\description{
Mask across tibble by column differences
}
