% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_micro_summary.R
\name{masking}
\alias{masking}
\title{Two-dimensional micro data masking to handle binary and nominal data}
\usage{
masking(
  body,
  tb = NULL,
  f1,
  ns,
  Ns,
  cut.off = cut.off,
  glue.mask = glue.mask,
  dec = dec,
  minimal = FALSE
)
}
\arguments{
\item{body}{full table body}

\item{tb}{filtered table body}

\item{f1}{Subsets indexes of relevant columns}

\item{ns}{Relevant ns arranged in matrix}

\item{Ns}{All Ns}

\item{cut.off}{micro cut point}

\item{glue.mask}{glue string for masking}

\item{dec}{decimals}

\item{minimal}{flag to mask minimally if no overall column is present}
}
\value{
tibble
}
\description{
Two-dimensional micro data masking to handle binary and nominal data
}
