% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Setup.R
\name{MBITES.Male.Setup}
\alias{MBITES.Male.Setup}
\title{Initialize Male Methods for M-BITES}
\usage{
MBITES.Male.Setup(overwrite = TRUE, timing = "exponential")
}
\arguments{
\item{timing}{character in "exponential", "gamma", "weibull"}
}
\value{
writes methods to \code{\link{MosquitoMale}} class.
}
\description{
This function initializes methods for M-BITES Male module; this function adds methods to the \code{\link{MosquitoMale}} class.
}
