% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Complex-Survival.R
\name{mbites_get_surviveFlightProb}
\alias{mbites_get_surviveFlightProb}
\title{M-BITES: Get Baseline Survival Probability for \code{\link{MosquitoFemale}}}
\usage{
mbites_get_surviveFlightProb()
}
\description{
Get baseline survival probability for \code{\link{mbitesGeneric_surviveFlight}}.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_surviveFlightProb()}.
}
}
