% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Complex-Bouts.R
\name{mbites_OvipositSearchCheck}
\alias{mbites_OvipositSearchCheck}
\title{M-BITES: Check for Oviposit Search Bout \code{MosquitoFemale}}
\usage{
mbites_OvipositSearchCheck()
}
\description{
During a resting bout \code{\link{mbites_boutR}}, the mosquito checks if she is in a
\itemize{
\item this method is bound to \code{MosquitoFemale$OvipositSearchCheck}
}
}
