% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_timingGamma}
\alias{mbitesBRO_timingGamma}
\title{MBITES-BRO: Gamma Timing for \code{MosquitoFemale}}
\usage{
mbitesBRO_timingGamma()
}
\description{
Method for Gamma-distributed bout lengths (model mosquitoes as a semi-Markov process).
\itemize{
\item This method is bound to \code{MosquitoFemale$timingExponential()}.
}
}
