% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Sugar.R
\name{mbitesBROS_boutS}
\alias{mbitesBROS_boutS}
\title{MBITES-BROS: Sugar Feeding Bout \code{MosquitoFemale}}
\usage{
mbitesBROS_boutS()
}
\description{
A mosquito performs a sugar feeding bout (all actions taken launch to launch when resting required).
Upon entering the sugar feeding behavioral state prompted by \code{\link{mbitesGeneric_queueSugarBout}}, the mosquito will move to a \code{\link{SugarSite}} and
sugar feed .
\itemize{
\item This method is bound to \code{MosquitoFemale$boutS()}.
}
}
