% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic-Energetics.R
\name{mbitesGeneric_flightEnergetics_Mean}
\alias{mbitesGeneric_flightEnergetics_Mean}
\title{MBITES-Generic: Mean Flight Energetics for \code{\link{MosquitoFemale}}}
\usage{
mbitesGeneric_flightEnergetics_Mean()
}
\description{
Reduce this mosquito's energy reserves by mean amount averaged over possible flight distances.
\itemize{
\item This method is bound to \code{MosquitoFemale$flightEnergetics()}.
}
}
