% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic-Energetics.R
\name{mbitesGeneric_queueSugarBout}
\alias{mbitesGeneric_queueSugarBout}
\title{MBITES-Generic: Queue Sugar Bout due to Energy Reserves for \code{\link{MosquitoFemale}}}
\usage{
mbitesGeneric_queueSugarBout()
}
\description{
Potentially queue a sugar bout \code{\link{mbitesBROS_boutS}} as a function of energy reserves.
\itemize{
\item This method is bound to \code{MosquitoFemale$queueSugarBout()}.
}
}
