% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic-Survival.R
\name{mbitesGeneric_WingTattering_Exact}
\alias{mbitesGeneric_WingTattering_Exact}
\title{MBITES-Generic: Exact Wing Tattering for \code{\link{MosquitoFemale}}}
\usage{
mbitesGeneric_WingTattering_Exact()
}
\description{
write me!
}
\details{
\itemize{
\item This method is bound to \code{MosquitoFemale$WingTattering()}.
}
}
