% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic-HostEncounter.R
\name{mbitesGeneric_zooEncounter}
\alias{mbitesGeneric_zooEncounter}
\title{MBITES-BRO: Non-Human Host Encounter for \code{\link{MosquitoFemale}}}
\usage{
mbitesGeneric_zooEncounter()
}
\description{
After calling \code{\link{mbitesGeneric_chooseHost}}, the mosquito encounters the non-human host and attempts to feed.
\itemize{
\item This method is bound to \code{MosquitoFemale$zooEncounter()}.
}
}
