% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_boutS}
\alias{mbitesMale_boutS}
\title{MBITES-Male: Male Sugar Feeding Bout \code{MosquitoMale}}
\usage{
mbitesMale_boutS()
}
\description{
A male mosquito attempts to sugar feed. If successful (governed by parameter S_succeed), the mosquito energy is topped off, otherwise the mosquito's energy is unchanged.
\itemize{
\item This method is bound to \code{MosquitoMale$boutS()}.
}
}
