% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_enterHouse}
\alias{mbitesMale_enterHouse}
\title{MBITES-Male: Attempt to Enter a House for \code{MosquitoMale}}
\usage{
mbitesMale_enterHouse()
}
\description{
Method to simulate attempted house entry for mosquito, and call appropriate events if the mosquito enters.
\itemize{
\item This method is bound to \code{MosquitoMale$enterHouse()}.
}
}
