% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Survival.R
\name{mbitesMale_get_restHaz}
\alias{mbitesMale_get_restHaz}
\title{MBITES-Male: Get Resting Hazards for \code{\link{MosquitoMale}}}
\usage{
mbitesMale_get_restHaz()
}
\description{
Get resting hazards for \code{\link{mbitesGeneric_surviveResting}}.
\itemize{
\item This method is bound to \code{MosquitoMale$get_restHaz()}.
}
}
