% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_oneMosquito_MBITES}
\alias{mbitesMale_oneMosquito_MBITES}
\title{MBITES-Male: Run Simulation for \code{\link{MosquitoMale}}}
\usage{
mbitesMale_oneMosquito_MBITES()
}
\description{
Run the M-BITES life cycle simulation algorithm while alive and has not overrun time in enclosing \code{\link{MicroTile}}.
This method calls \code{\link{mbitesMale_oneBout}} to simulate each life stage.
\itemize{
\item This method is bound to \code{MosquitoMale$MBITES()}.
}
}
