% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Energetics.R
\name{mbitesMale_pEnergySurvival}
\alias{mbitesMale_pEnergySurvival}
\title{MBITES-Male: Probability of Death due to Energy Reserves for \code{\link{MosquitoMale}}}
\usage{
mbitesMale_pEnergySurvival()
}
\description{
Calculate mortality probability as a function of energy reserves.
\itemize{
\item This method is bound to \code{MosquitoMale$pEnergySurvival()}.
}
}
