% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Energetics.R
\name{mbitesMale_queueSugarBout}
\alias{mbitesMale_queueSugarBout}
\title{MBITES-Male: Queue Sugar Bout due to Energy Reserves for \code{\link{MosquitoMale}}}
\usage{
mbitesMale_queueSugarBout()
}
\description{
Potentially queue a sugar bout \code{\link{mbitesMale_boutS}} as a function of energy reserves.
\itemize{
\item This method is bound to \code{MosquitoMale$queueSugarBout()}.
}
}
