% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Energetics.R
\name{mbitesMale_sugarEnergetics}
\alias{mbitesMale_sugarEnergetics}
\title{MBITES-Male: Sugar Energetics for \code{\link{MosquitoMale}}}
\usage{
mbitesMale_sugarEnergetics()
}
\description{
Handle energy dependent mortality and sugar bout queueing as function of current mosquito energy levels.
\itemize{
\item This method is bound to \code{MosquitoMale$sugarEnergetics()}.
}
}
