% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Survival.R
\name{mbitesMale_surviveFlight}
\alias{mbitesMale_surviveFlight}
\title{MBITES-Male: Flight Survival for \code{\link{MosquitoMale}}}
\usage{
mbitesMale_surviveFlight()
}
\description{
Run generic flight survival probailities for bouts (launch to launch).
Depending on settings from M-BITES parameters, senescence and/or tattering may also be simulated.
\itemize{
\item This method is bound to \code{MosquitoMale$surviveFlight()}.
}
}
