% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Survival.R
\name{mbitesMale_surviveResting}
\alias{mbitesMale_surviveResting}
\title{MBITES-Male: Resting Survival for \code{\link{MosquitoMale}}}
\usage{
mbitesMale_surviveResting()
}
\description{
Run generic resting survival probailities for bouts (launch to launch).
\itemize{
\item This method is bound to \code{MosquitoMale$surviveResting()}.
}
}
