% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Survival.R
\name{mbitesMale_WingTattering}
\alias{mbitesMale_WingTattering}
\title{MBITES-Generic: Wing Tattering for \code{\link{MosquitoFemale}}}
\usage{
mbitesMale_WingTattering()
}
\description{
Draw from a zero-inflated Beta distribution for additive wing damage from tattering.
Wing damage is given by \deqn{ \left\{\begin{matrix}
x=0; P(ttsz.p)
\\
x\sim Beta(ttsz.a,ttsz.b); P(1-ttsz.p)
\end{matrix}\right. }
\itemize{
\item This method is bound to \code{MosquitoMale$WingTattering()}.
}
}
