% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel-Generate.R
\name{MicroKernel_DistanceMat}
\alias{MicroKernel_DistanceMat}
\title{MICRO Search Kernels: Calculate Euclidean Distance Matrix}
\usage{
MicroKernel_DistanceMat(S, D)
}
\arguments{
\item{S}{pointer to list of source sites (for example, Landscape$get_AquaSites())}

\item{D}{pointer to list of destination sites}
}
\value{
numeric matrix
}
\description{
Given pointers \code{S} (source) and \code{D} (destination) to two lists of \code{Site} objects,
Calculate the Euclidean distance matrix between them.
}
