% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel-Methods.R
\name{MicroKernel_SampleMvOb_MosquitoPopFemale}
\alias{MicroKernel_SampleMvOb_MosquitoPopFemale}
\title{MICRO Search Kernels: \code{\link{MosquitoPopFemale}} Sample MvOb}
\usage{
MicroKernel_SampleMvOb_MosquitoPopFemale(locNow, state, pSetNow)
}
\arguments{
\item{locNow:}{current site index \code{ix} of this mosquito}

\item{state:}{current behavioral state of this mosquito}

\item{inPointSet:}{current point set this mosquito is in}
}
\value{
new index
}
\description{
This method is a helper for \code{\link{MicroKernel_moveMe}} and samples the appropriate MvOb in MvAll stored in an enclosing \code{\link{Tile}} object.
\itemize{
\item This function is bound to \code{MosquitoPopFemale$SampleMove}
}
}
