% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsim.R
\name{microsim}
\alias{microsim}
\title{Demographic Microsimulation}
\usage{
microsim(
  initial_popn,
  fertility_rates = NULL,
  mortality_rates,
  internal_rates = NULL,
  immigration_means,
  emigration_rates,
  step_length,
  dominant_sex = "Female"
)
}
\arguments{
\item{initial_popn}{An array with dimensions "age",
"sex", and "region"}

\item{fertility_rates}{An array with dimensions
"age", "triangle", "sex", "region", and "time".
Optional.}

\item{mortality_rates}{An array with dimensions
"age", "triangle", "sex", "region", and "time".}

\item{internal_rates}{An array with dimensions
"age", "triangle", "sex", "region_orig",
"region_dest", and "time". Optional.}

\item{immigration_means}{An array with dimensions
"age", "triangle", "sex", "region", and "time".}

\item{emigration_rates}{An array with dimensions
"age", "triangle", "sex", "region", and "time".}

\item{step_length}{Age-time steps. Usually 1 or 5.}

\item{dominant_sex}{The sex to which births are
attributed. Defaults to \code{"Female"}.}
}
\value{
An object of class \code{\link[dembase]{Movements-class}}.
}
\description{
Given a set of demographic rates, plus an initial population,
use microsimulation to create a demographic account.
The microsimulation involves random draws, so the
demographic account is probabilistic.
}
\details{
The interface for the function is not fully developed,
and the function is a little inflexible. In due course,
the function (and package 'micro') will be superceded by
a more general version.

All dimensions with the same name should have the same length:
for instance, all dimensions called "age" should have the same
length. This includes the "age" dimension of \code{fertility_rates}.
Rates outside the reproductive ages (normally 15 to 49)
should be set to zero.

The "triangle" dimension holds Lexis triangles, and
must have labels "Lower" and "Upper".

The "sex" dimension can be length 1, 2, or more.

The "region" dimension is required, but can be length 1.

\code{initial_popn} is population counts, \code{fertility_rates},
\code{mortality_rates}, \code{internal_rates},
and \code{emigration_rates} are all rates,
and \code{immigration_means} is expected values.
}
