% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_micro_table.R
\name{minimal_mask}
\alias{minimal_mask}
\title{Small function to mask minimally}
\usage{
minimal_mask(mask, data, micro.n)
}
\arguments{
\item{mask}{matrix of logical masking}

\item{data}{tibble}

\item{micro.n}{n to mask at}
}
\value{
tibble
}
\description{
Small function to mask minimally
}
\examples{
ls <- list(mask = tibble::as_tibble(matrix(c(FALSE, FALSE, TRUE, TRUE), nrow = 1),
    .name_repair = "unique_quiet"
  ),
  data = tibble::as_tibble(matrix(c(20, 12, 8, 4), nrow = 1),
    .name_repair = "unique_quiet"
  ), micro.n = 5
)
# minimal_mask(ls$mask,ls$data,5)
}
