% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Class.R
\docType{class}
\name{MosquitoFemale}
\alias{MosquitoFemale}
\title{MosquitoFemale Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
MosquitoFemale
}
\description{
im a class!
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item id: character of format tEmerge_id_genotype (ie; "5_42_2" is the 42nd mosquito emerging on day 5 with genotype 2)
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
