% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{oneDay_EL4P}
\alias{oneDay_EL4P}
\title{EL4P Daily Difference Equations}
\usage{
oneDay_EL4P()
}
\description{
Run daily EL4P difference equations.
\itemize{
\item This method is bound to \code{EL4P$oneDay_EL4P}
}
}
