% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Emerge-Methods.R
\name{oneStep_EmergeSite}
\alias{oneStep_EmergeSite}
\title{MICRO \code{\link{AquaticSite}} Method: Emerge One Day Dynamics}
\usage{
oneStep_EmergeSite(tNow)
}
\arguments{
\item{tNow}{current global time of \code{\link{MicroTile}}}
}
\description{
Calculate emerging adults for a single aquatic habitat and add them to that site's ImagoQ.
\itemize{
\item This method is bound to \code{AquaticSite$oneStep_EmergeSite} and \code{periDomestic_AquaticSite$oneStep_EmergeSite}
}
}
