% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-PeriDomestic.R
\docType{class}
\name{periDomestic_AquaticSite}
\alias{periDomestic_AquaticSite}
\title{Peri-Domestic Breeding Site Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
periDomestic_AquaticSite
}
\description{
A peri-domestic breeding site is an aquatic habitat that is considered to be
part of an enclosing \code{\link{FeedingSite}}, for example a backyard pool or water container.
If a mosquito is at a feeding site with a peri-domestic breeding site, she does not need to enter a full
search bout to oviposit.
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
