% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHOGEN-PfSI-Setup.R
\docType{data}
\name{PfSI_NULL}
\alias{PfSI_NULL}
\title{PfSI \code{NULL} Mosquito Pathogen History}
\format{An object of class \code{list} of length 1.}
\usage{
PfSI_NULL
}
\description{
This is used to pad the final entry when mosquito-stage pathogen histories are being written to JSON by \code{\link[jsonlite]{toJSON}}.
}
\keyword{datasets}
