% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Simulation.R
\name{resetMicro}
\alias{resetMicro}
\title{Reset Tile Between Runs}
\usage{
resetMicro(MosquitoPar, HumanPar, EL4P = FALSE, mating = FALSE)
}
\description{
write me
}
\details{
\itemize{
\item This method is bound to \code{Tile$resetMicro}
}
}
