% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{rmSelf_MosquitoFemale}
\alias{rmSelf_MosquitoFemale}
\title{Remove Self}
\usage{
rmSelf_MosquitoFemale()
}
\description{
Remove this mosquito from the population containing it.
\itemize{
\item This method is bound to \code{MosquitoFemale$rmSelf}
}
}
