% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{rmSelf_MosquitoMale}
\alias{rmSelf_MosquitoMale}
\title{Remove Self}
\usage{
rmSelf_MosquitoMale()
}
\description{
Remove this mosquito from the population containing it.
\itemize{
\item This method is bound to \code{MosquitoMale$rmSelf}
}
}
