% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounded_interval.R
\name{rounded_interval}
\alias{rounded_interval}
\title{Create interval with value rounded down and up to nearest number specified}
\usage{
rounded_interval(data, round = 5, down = TRUE)
}
\arguments{
\item{data}{numeric vector}

\item{round}{number to round to}

\item{down}{round down on tie? Default is TRUE. Rounds up if FALSE.}
}
\value{
vector
}
\description{
Create interval with value rounded down and up to nearest number specified
}
\examples{
1:10 |> rounded_interval(round=2)
}
