% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-ActivitySpace-Methods.R
\name{set_ActivitySpace_Human}
\alias{set_ActivitySpace_Human}
\title{Set \code{\link[MASHmacro]{Human}} ActivitySpace}
\usage{
set_ActivitySpace_Human(nDaily, Nplaces, p, loc)
}
\arguments{
\item{nDaily}{integer average daily number of other sites visited}

\item{Nplaces}{integer number of places visited}

\item{p}{weight numeric on proportion of time spent at home}

\item{loc}{integer vector of other sites visited}
}
\description{
This function is bound to \code{Human$set_ActivitySpace()}
}
