% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods-VectorControl.R
\name{set_attractiveSugarBait}
\alias{set_attractiveSugarBait}
\title{Set attractiveSugarBait}
\usage{
set_attractiveSugarBait(atsbDefinition)
}
\arguments{
\item{attractiveSugarBait}{Object}
}
\description{
Change the definition of the ATSB in the sugar site
}
