% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{set_directory_Tile}
\alias{set_directory_Tile}
\title{Set Directory}
\usage{
set_directory_Tile(directory)
}
\arguments{
\item{directory}{character}
}
\description{
Set the directory (character) containing output for this tile
\itemize{
\item This method is bound to \code{Tile$set_directory}
}
}
