% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-PeriDomestic.R
\name{set_EggQ_periDomestic_AquaticSite}
\alias{set_EggQ_periDomestic_AquaticSite}
\title{Set EggQ}
\usage{
set_EggQ_periDomestic_AquaticSite(EggQ)
}
\arguments{
\item{EggQ}{object of class \code{\link[MASHcpp]{EggQ}}}
}
\description{
Set mosquito egg batch queue, object of class \code{\link[MASHcpp]{EggQ}}
\itemize{
\item This method is bound to \code{periDomestic_AquaticSite$set_EggQ}
}
}
