% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_EL4P_AquaticSite}
\alias{set_EL4P_AquaticSite}
\title{Set EL4P}
\usage{
set_EL4P_AquaticSite(EL4P)
}
\arguments{
\item{EL4P}{object of class \code{\link[MASHcpp]{EL4P}}}
}
\description{
Set aquatic populations, object of class \code{\link[MASHcpp]{EL4P}}
\itemize{
\item This method is bound to \code{AquaticSite$set_EL4P}
}
}
