% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{set_FeedingSites_Landscape}
\alias{set_FeedingSites_Landscape}
\title{Set Blood Feeding Sites}
\usage{
set_FeedingSites_Landscape(FeedingSites, ixS = NULL)
}
\arguments{
\item{FeedingSites}{input may be a single \code{\link{FeedingSite}} or list of sites}

\item{ixS}{integer index; if \code{NULL} set entire list of sites}
}
\description{
Set a single \code{\link{FeedingSite}} or all sites.
\itemize{
\item This method is bound to \code{Landscape$set_FeedingSites}
}
}
