% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{set_FemaleCSVCon_Tile}
\alias{set_FemaleCSVCon_Tile}
\title{Set Female CSV Connection}
\usage{
set_FemaleCSVCon_Tile(FemaleCSVCon)
}
\arguments{
\item{FemaleCSVCon}{object of class \code{\link[base]{connection}}, usually opened with \code{file}}
}
\description{
Set the text connection used to write population count data to CSV.
\itemize{
\item This method is bound to \code{Tile$set_FemaleCSVCon}
}
}
