% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{set_FemalePopPointer_Landscape}
\alias{set_FemalePopPointer_Landscape}
\title{Set \code{\link{MosquitoPopFemale}} Pointer}
\usage{
set_FemalePopPointer_Landscape(FemalePopPointer)
}
\arguments{
\item{FemalePopPointer}{an environment}
}
\description{
Set pointer to \code{MosquitoPopFemale$self}.
\itemize{
\item This method is bound to \code{Landscape$set_FemalePopPointer}
}
}
