% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_FemalePopPointer_Mosquito}
\alias{set_FemalePopPointer_Mosquito}
\title{Set \code{\link{MosquitoPopFemale}} Pointer}
\usage{
set_FemalePopPointer_Mosquito(FemalePopPointer)
}
\arguments{
\item{FemalePopPointer}{an environment}
}
\description{
Set the pointer to \code{MosquitoPopFemale$self} enclosing this mosquito (if method for \code{\link{MosquitoFemale}}), or in the same \code{\link{Tile}} (if method for \code{\link{MosquitoMale}}).
\itemize{
\item This method is bound to \code{MosquitoFemale$set_FemalePopPointer} and \code{MosquitoMale$set_FemalePopPointer}
}
}
