% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{set_FemalePopPointer_MosquitoPopMale}
\alias{set_FemalePopPointer_MosquitoPopMale}
\title{Set \code{\link{MosquitoPopFemale}} Pointer}
\usage{
set_FemalePopPointer_MosquitoPopMale(FemalePopPointer)
}
\arguments{
\item{MalePopPointer}{an environment}
}
\description{
Set the pointer to \code{MosquitoPopMale$self} in the same \code{\link{Tile}} by reference.
\itemize{
\item This method is bound to \code{MosquitoPopMale$set_FemalePopPointer}
}
}
