% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_haz_MatingSite}
\alias{set_haz_MatingSite}
\title{Set haz}
\usage{
set_haz_MatingSite(haz)
}
\arguments{
\item{haz}{numeric}
}
\description{
Set landing hazard associated with outside wall at this site \code{haz}
\itemize{
\item This method is bound to \code{MatingSite$set_haz}
}
}
