% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_hazI_FeedingSite}
\alias{set_hazI_FeedingSite}
\title{Set hazI}
\usage{
set_hazI_FeedingSite(hazI)
}
\arguments{
\item{hazI}{numeric}
}
\description{
Set landing hazard associated with indoor wall at this site \code{hazI}
\itemize{
\item This method is bound to \code{FeedingSite$set_hazI}
}
}
