% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_hazV_FeedingSite}
\alias{set_hazV_FeedingSite}
\title{Set hazV}
\usage{
set_hazV_FeedingSite(hazV)
}
\arguments{
\item{hazV}{numeric}
}
\description{
Set landing hazard associated with vegetation at this site \code{hazV}
\itemize{
\item This method is bound to \code{FeedingSite$set_hazV}
}
}
