% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_hazW_FeedingSite}
\alias{set_hazW_FeedingSite}
\title{Set hazW}
\usage{
set_hazW_FeedingSite(hazW)
}
\arguments{
\item{hazW}{numeric}
}
\description{
Set landing hazard associated with outside wall at this site \code{hazW}
\itemize{
\item This method is bound to \code{FeedingSite$set_hazW}
}
}
