% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{set_HumanPathogenCon_Tile}
\alias{set_HumanPathogenCon_Tile}
\title{Set Human Pathogen JSON Connection}
\usage{
set_HumanPathogenCon_Tile(HumanPathogenCon)
}
\arguments{
\item{HumanPathogenCon}{object of class \code{\link[base]{connection}}, usually opened with \code{file}}
}
\description{
Set the text connection used to write human-stage pathogen data to JSON
\itemize{
\item This method is bound to \code{Tile$set_HumanPathogenCon}
}
}
