% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_HumansPointer_Mosquito}
\alias{set_HumansPointer_Mosquito}
\title{Set \code{\link[MASHmacro]{HumanPop}} Pointer}
\usage{
set_HumansPointer_Mosquito(HumansPointer)
}
\arguments{
\item{HumansPointer}{an environment}
}
\description{
Set pointer to \code{HumanPop$self} within the same microsimulation tile.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_HumansPointer} and \code{MosquitoMale$set_HumansPointer}
}
}
