% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{set_HumansPointer_MosquitoPop}
\alias{set_HumansPointer_MosquitoPop}
\title{Set \code{\link[MASHmacro]{HumanPop}} Pointer}
\usage{
set_HumansPointer_MosquitoPop(HumansPointer)
}
\arguments{
\item{HumansPointer}{an environment}
}
\description{
Set pointer to \code{HumanPop$self} within the same microsimulation tile.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$set_HumansPointer} and \code{MosquitoPopMale$set_HumansPointer}
}
}
