% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_ImagoQ_AquaticSite}
\alias{set_ImagoQ_AquaticSite}
\title{Set ImagoQ}
\usage{
set_ImagoQ_AquaticSite(ImagoQ)
}
\arguments{
\item{ImagoQ}{object of class \code{\link[MASHcpp]{ImagoQ}}}
}
\description{
Set mosquito imago queue of emerging adults, object of class \code{\link[MASHcpp]{ImagoQ}}
\itemize{
\item This method is bound to \code{AquaticSite$set_ImagoQ}
}
}
