% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods-VectorControl.R
\name{set_indoorResidualSpray}
\alias{set_indoorResidualSpray}
\title{Set swarmSpray}
\usage{
set_indoorResidualSpray(indoorResidualSprayDefinition)
}
\arguments{
\item{swarmSpray}{Object}
}
\description{
Change the definition of the Aerial Spray
}
