% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_ix_SugarSite}
\alias{set_ix_SugarSite}
\title{Set ix}
\usage{
set_ix_SugarSite(ix)
}
\arguments{
\item{ix}{integer}
}
\description{
Set site index \code{ix}
\itemize{
\item This method is bound to \code{SugarSite$set_ix}
}
}
