% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_L2_EL4P}
\alias{set_L2_EL4P}
\title{EL4P: Set L2}
\usage{
set_L2_EL4P(L2)
}
\arguments{
\item{L2}{numeric}
}
\description{
Set L2
\itemize{
\item This method is bound to \code{EL4P$set_L2}
}
}
