% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_L3_EL4P}
\alias{set_L3_EL4P}
\title{EL4P: Set L3}
\usage{
set_L3_EL4P(L3)
}
\arguments{
\item{L3}{numeric}
}
\description{
Set L3
\itemize{
\item This method is bound to \code{EL4P$set_L3}
}
}
