% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_L4_EL4P}
\alias{set_L4_EL4P}
\title{EL4P: Set L4}
\usage{
set_L4_EL4P(L4)
}
\arguments{
\item{L4}{numeric}
}
\description{
Set L4
\itemize{
\item This method is bound to \code{EL4P$set_L4}
}
}
