% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_lambda_EL4P}
\alias{set_lambda_EL4P}
\title{EL4P: Set Lambda}
\usage{
set_lambda_EL4P(lambda)
}
\arguments{
\item{lambda}{numeric}
}
\description{
Set lambda
\itemize{
\item This method is bound to \code{EL4P$set_lambda}
}
}
