% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_LandscapePointer_FeedingSite}
\alias{set_LandscapePointer_FeedingSite}
\title{Set \code{\link{Landscape}} Pointer}
\usage{
set_LandscapePointer_FeedingSite(LandscapePointer)
}
\arguments{
\item{LandscapePointer}{an environment}
}
\description{
Set the pointer \code{Landscape$self} enclosing this site.
\itemize{
\item This method is bound to \code{FeedingSite$set_LandscapePointer}
}
}
