% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Setup.R
\name{set_LandscapePointer_HumanPop}
\alias{set_LandscapePointer_HumanPop}
\title{Set \code{\link{Landscape}} Pointer}
\usage{
set_LandscapePointer_HumanPop(LandscapePointer)
}
\arguments{
\item{LandscapePointer}{an environment}
}
\description{
Set pointer to \code{Landscape$self} this population lives in.
\itemize{
\item This method is bound to \code{HumanPop$set_LandscapePointer}
}
}
