% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_LandscapePointer_Mosquito}
\alias{set_LandscapePointer_Mosquito}
\title{Set \code{\link{Landscape}} Pointer}
\usage{
set_LandscapePointer_Mosquito(LandscapePointer)
}
\arguments{
\item{LandscapePointer}{an environment}
}
\description{
Set pointer to the enclosing \code{Landscape$self}.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_LandscapePointer} and \code{MosquitoMale$set_LandscapePointer}
}
}
