% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{set_LandscapePointer_MosquitoPop}
\alias{set_LandscapePointer_MosquitoPop}
\title{Set \code{\link{Landscape}} Pointer}
\usage{
set_LandscapePointer_MosquitoPop(LandscapePointer)
}
\arguments{
\item{LandscapePointer}{an environment}
}
\description{
Set pointer to the enclosing \code{Landscape$self}.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$set_LandscapePointer} and \code{MosquitoPopMale$set_LandscapePointer}
}
}
