% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods-VectorControl.R
\name{set_larviciding}
\alias{set_larviciding}
\title{Set swarmSpray}
\usage{
set_larviciding(larvicidingDefinition)
}
\arguments{
\item{swarmSpray}{Object}
}
\description{
Change the definition of the Aerial Spray
}
