% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_locNow_Mosquito}
\alias{set_locNow_Mosquito}
\title{Set locNow}
\usage{
set_locNow_Mosquito(loc)
}
\arguments{
\item{loc}{character}
}
\description{
Set current mosquito point set (\code{character}). With \code{pSetNow}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_locNow} and \code{MosquitoMale$set_locNow}
}
}
