% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Setup.R
\name{set_MalePopPointer_HumanPop}
\alias{set_MalePopPointer_HumanPop}
\title{Set \code{\link{MosquitoPopMale}} Pointer}
\usage{
set_MalePopPointer_HumanPop(MalePopPointer)
}
\arguments{
\item{MosquitoPopMale}{an environment}
}
\description{
Set pointer to the female mosquito population on this landscape.
\itemize{
\item This method is bound to \code{HumanPop$set_MalePopPointer}
}
}
