% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{set_MalePopPointer_Landscape}
\alias{set_MalePopPointer_Landscape}
\title{Set \code{\link{MosquitoPopMale}} Pointer}
\usage{
set_MalePopPointer_Landscape(MalePopPointer)
}
\arguments{
\item{MalePopPointer}{an environment}
}
\description{
Set pointer to \code{MosquitoPopFemale$self}.
\itemize{
\item This method is bound to \code{Landscape$set_MalePopPointer}
}
}
