% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_MalePopPointer_Mosquito}
\alias{set_MalePopPointer_Mosquito}
\title{Set \code{\link{MosquitoPopMale}} Pointer}
\usage{
set_MalePopPointer_Mosquito(MalePopPointer)
}
\arguments{
\item{MalePopPointer}{an environment}
}
\description{
Set the pointer to \code{MosquitoPopMale$self} enclosing this mosquito (if method for \code{\link{MosquitoMale}}), or in the same \code{\link{Tile}} (if method for \code{\link{MosquitoFemale}}).
\itemize{
\item This method is bound to \code{MosquitoFemale$set_MalePopPointer} and \code{MosquitoMale$set_MalePopPointer}
}
}
