% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{set_MalePopPointer_MosquitoPopFemale}
\alias{set_MalePopPointer_MosquitoPopFemale}
\title{Set \code{\link{MosquitoPopMale}} Pointer}
\usage{
set_MalePopPointer_MosquitoPopFemale(MalePopPointer)
}
\arguments{
\item{MalePopPointer}{an environment}
}
\description{
Set the pointer to \code{MosquitoPopMale$self} in the same \code{\link{Tile}} by reference.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$set_MalePopPointer}
}
}
