% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{set_MatingSites_Landscape}
\alias{set_MatingSites_Landscape}
\title{Set Mating Sites}
\usage{
set_MatingSites_Landscape(MatingSites, ixS = NULL)
}
\arguments{
\item{MatingSites}{input may be a single \code{\link{MatingSite}} or list of sites}

\item{ixS}{integer index; if \code{NULL} set entire list of sites}
}
\description{
Set a single \code{\link{MatingSite}} or all sites.
\itemize{
\item This method is bound to \code{Landscape$set_MatingSites}
}
}
