% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{set_MosquitoPathogenCon_Tile}
\alias{set_MosquitoPathogenCon_Tile}
\title{Set Mosquito Pathogen JSON Connection}
\usage{
set_MosquitoPathogenCon_Tile(MosquitoPathogenCon)
}
\arguments{
\item{MosquitoPathogenCon}{object of class \code{\link[base]{connection}}, usually opened with \code{file}}
}
\description{
Set the text connection used to write mosquito-stage pathogen data to JSON
\itemize{
\item This method is bound to \code{Tile$set_MosquitoPathogenCon}
}
}
