% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_mu_E_EL4P}
\alias{set_mu_E_EL4P}
\title{EL4P: Set mu_E}
\usage{
set_mu_E_EL4P(mu_E)
}
\arguments{
\item{mu_E}{numeric}
}
\description{
Set mu_E
\itemize{
\item This method is bound to \code{EL4P$set_mu_E}
}
}
