% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_mu_L_EL4P}
\alias{set_mu_L_EL4P}
\title{EL4P: Set mu_L}
\usage{
set_mu_L_EL4P(mu_L)
}
\arguments{
\item{mu_L}{numeric}
}
\description{
Set mu_L
\itemize{
\item This method is bound to \code{EL4P$set_mu_L}
}
}
