% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_mu_P_EL4P}
\alias{set_mu_P_EL4P}
\title{EL4P: Set mu_P}
\usage{
set_mu_P_EL4P(mu_P)
}
\arguments{
\item{mu_P}{numeric}
}
\description{
Set mu_P
\itemize{
\item This method is bound to \code{EL4P$set_mu_P}
}
}
