% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{set_runID_Tile}
\alias{set_runID_Tile}
\title{Set Run ID}
\usage{
set_runID_Tile(runID)
}
\arguments{
\item{runID}{integer}
}
\description{
Set current simulation run ID that will be used to identify multiple simulation output from the same tile object.
\itemize{
\item This method is bound to \code{Tile$set_runID}
}
}
