% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_searchWt_MatingSite}
\alias{set_searchWt_MatingSite}
\title{Set searchWt}
\usage{
set_searchWt_MatingSite(searchWt)
}
\arguments{
\item{ix}{numeric}
}
\description{
Set site search weight \code{searchWt}
\itemize{
\item This method is bound to \code{MatingSite$set_searchWt}
}
}
