% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_siteType_FeedingSite}
\alias{set_siteType_FeedingSite}
\title{Set siteType}
\usage{
set_siteType_FeedingSite(siteType)
}
\arguments{
\item{siteType}{integer}
}
\description{
Set site type \code{siteType}; 1 is peri-domestic, 0 is non peri-domestic.
\itemize{
\item This method is bound to \code{FeedingSite$set_siteType}
}
}
