% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_siteXY_MatingSite}
\alias{set_siteXY_MatingSite}
\title{Set siteXY}
\usage{
set_siteXY_MatingSite(newSiteXY)
}
\arguments{
\item{ix}{numeric(2)}
}
\description{
Set site coordinates \code{siteXY}
\itemize{
\item This method is bound to \code{MatingSite$set_siteXY}
}
}
