% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{set_SugarSites_Landscape}
\alias{set_SugarSites_Landscape}
\title{Set Sugar Feeding Sites}
\usage{
set_SugarSites_Landscape(SugarSites, ixS = NULL)
}
\arguments{
\item{SugarSites}{input may be a single \code{\link{SugarSite}} or list of sites}

\item{ixS}{integer index; if \code{NULL} set entire list of sites}
}
\description{
Set a single \code{\link{SugarSite}} or all sites.
\itemize{
\item This method is bound to \code{Landscape$set_SugarSites}
}
}
