% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{set_TilePointer_Landscape}
\alias{set_TilePointer_Landscape}
\title{Set \code{\link{Tile}} Pointer}
\usage{
set_TilePointer_Landscape(TilePointer)
}
\arguments{
\item{TilePointer}{an environment}
}
\description{
Set pointer to enclosing \code{Tile$self}.
\itemize{
\item This method is bound to \code{Landscape$set_TilePointer}
}
}
