% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_TilePointer_Mosquito}
\alias{set_TilePointer_Mosquito}
\title{Set \code{\link{Tile}} Pointer}
\usage{
set_TilePointer_Mosquito(TilePointer)
}
\arguments{
\item{TilePointer}{an environment}
}
\description{
Set pointer to enclosing \code{Tile$self}.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_TilePointer} and \code{MosquitoMale$set_TilePointer}
}
}
