% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{set_TilePointer_MosquitoPop}
\alias{set_TilePointer_MosquitoPop}
\title{Set \code{\link{Tile}} Pointer}
\usage{
set_TilePointer_MosquitoPop(TilePointer)
}
\arguments{
\item{TilePointer}{an environment}
}
\description{
Set pointer to enclosing \code{Tile$self}.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$set_TilePointer} and \code{MosquitoPopMale$set_TilePointer}
}
}
