% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{set_tNow_Tile}
\alias{set_tNow_Tile}
\title{Set tNow}
\usage{
set_tNow_Tile(tNow)
}
\arguments{
\item{tNow}{numeric}
}
\description{
Set current simulation time
\itemize{
\item This method is bound to \code{Tile$set_tNow}
}
}
