% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-ActivitySpace-Methods.R
\name{sim_ActivitySpace_Human}
\alias{sim_ActivitySpace_Human}
\title{MICRO \code{\link{Human}} Method: Initialize Activity Space}
\usage{
sim_ActivitySpace_Human()
}
\description{
This function is bound to \code{Human$sim_ActivitySpace()}
}
