% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_micro_summary.R
\name{summary_masks}
\alias{summary_masks}
\title{Implementation step for masking}
\usage{
summary_masks(body, ls, cut.off = 5, dec = dec)
}
\arguments{
\item{body}{table body}

\item{ls}{complete summary table list object}

\item{cut.off}{micro cut point}

\item{dec}{decimals, passed on}
}
\value{
tibble
}
\description{
Implementation step for masking
}
