% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEBUG.R
\name{test_WriteHistoryAndDelete_MosquitoPopFemale}
\alias{test_WriteHistoryAndDelete_MosquitoPopFemale}
\title{Debug: Write Out Female Histories and Delete}
\usage{
test_WriteHistoryAndDelete_MosquitoPopFemale()
}
\description{
Testing writing out all female histories to JSON and then having each female delete herself from the MASHcpp::HashMap class.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$test_WriteHistoryAndDelete}
}
}
