% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-ActivitySpace-Methods.R
\name{write_ActivitySpace_Tile}
\alias{write_ActivitySpace_Tile}
\title{Write Human Activity Space on a Tile}
\usage{
write_ActivitySpace_Tile()
}
\description{
Write the human activity space for each site to JSON.
\itemize{
\item This function is bound to \code{Tile$write_ActivitySpace()}
}
}
