% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{writeAndDelete_MosquitoMale}
\alias{writeAndDelete_MosquitoMale}
\title{Export JSON History and Remove Self}
\usage{
writeAndDelete_MosquitoMale(conHist)
}
\arguments{
\item{conHist}{a \code{\link[base]{connection}} to write mosquito history to (the correct directory can be found with \code{\link{get_MosquitoDirectory_Tile}})}
}
\description{
\itemize{
\item This method is bound to \code{MosquitoMale$writeAndDelete}
}
}
